copy paste pdm related files
