+++ pycross/private/tools/BUILD.bazel
@@ -107,25 +107,15 @@ py_binary(
     imports = ["../../.."],
     visibility = ["//visibility:public"],
     deps = [
+        ":args",
         ":lock_model",
-        ":translation_utils",
+        "@rules_pycross_internal//deps:tomli",
     ],
 )
 
 py_binary(
     name = "uv_translator",
     srcs = ["uv_translator.py"],
-    imports = ["../../.."],
-    visibility = ["//visibility:public"],
-    deps = [
-        ":lock_model",
-        ":translation_utils",
-    ],
-)
-
-py_library(
-    name = "translation_utils",
-    srcs = ["translation_utils.py"],
     imports = ["../../.."],
     visibility = ["//visibility:public"],
     deps = [
